Added power managment
Add power management capability through DAPM. Pin Bias is assumed to control the enable of the IV stage opamp and GPIO2 is assumed to control the power supply of the opamp. When a music is about to start, it will first enable the power, wait 100ms, then start the opamp.